+++ commands/serve/web/eHub.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as ReactDOM from 'react-dom/client';
 
-import Hub from './components/Hub';
+import Hub from './components/Hub/Hub';
 
 ReactDOM.createRoot(document.querySelector('#hub')).render(<Hub />);
